게시글,댓글 수정,삭제시 sql로유저 idx 검사/ 세션,환경설정 server.js에서만 설정

+++ src/routers/article.js
@@ -1,39 +1,28 @@
 const router = require("express").Router();
-const session = require("express-session");
-// mysql db 연결,쿼리실행 모듈 import
-const mysql = require("mysql");
-// db세부정보 import
+const mysql = require('mysql');
 const dbconfig = require('../../config/db.js');
-require('dotenv').config();
-const secretCode = process.env.secretCode;
-router.use(session({
-    resave: false,
-    saveUninitialized: true,
-    secret: secretCode,
-    cookie: {
-        maxAge: 5 * 60 * 1000,
-        rolling:true
-    }
-}));
 
 //게시글 목록 불러오기route
 router.get("/all", (req, res) => {
     const result = {
         "success": false,
         "message": "실패",
-        "article": []
+        "data":{
+            "article": [] //data로
+        }
     }
     try {
         const conn = mysql.createConnection(dbconfig);
-        const sql = "SELECT a.idx, title, write_date, u.name FROM article a JOIN account u ON a.user_idx = u.idx ORDER BY write_date";
-        conn.query(sql, (err, rs) => {
+        const sql = "SELECT a.idx, title, write_date, u.name FROM article a JOIN account u ON a.user_idx = u.idx ORDER BY write_date"; //orderby는 idx로하기!
+        conn.query(sql, (err, rs) => { //rs는 원래 리스트
             try {
                 if (err) throw new Error("db 에러");
                 if (rs.length == 0) throw new Error("게시글없음");
-                rs.forEach((elem) => {
-                    let data = [elem.idx, elem.title, elem.write_date, elem.name];
-                    result.article.push(data);
-                })
+                // rs.forEach((elem) => {
+                //     let data = [elem.idx, elem.title, elem.write_date, elem.name];
+                //     result.article.push(data);
+                // })
+                result.article = rs;
             } catch (e) {
                 result.message = e.message;
             } finally {
@@ -56,7 +45,7 @@ router.post("/", (req, res) => {
         "message": "실패"
     }
     try {
-        //if (!req.session.idx) throw new Error("세션없음");
+        if (!req.session.idx) throw new Error("세션없음");
         const conn = mysql.createConnection(dbconfig);
         const sql = "INSERT INTO article(title,content,user_idx) VALUES (?,?,?) ";
         const values = [title, content, userIdx];
@@ -88,11 +77,10 @@ router.put("/:articleIdx", (req, res) => {
         "message": "실패"
     }
     try {
-        //if (req.session.idx !== userIdx) throw new Error("사용자 idx가 불일치");
-
+        if (req.session.idx !== userIdx) throw new Error("사용자 idx가 불일치");
         const conn = mysql.createConnection(dbconfig);
-        const sql = "UPDATE article SET title = ?, content = ? WHERE idx = ?";
-        const values = [title, content, articleIdx];
+        const sql = "UPDATE article SET title = ?, content = ? WHERE idx = ? AND user_idx = ?";
+        const values = [title, content, articleIdx, userIdx];
         //db통신
         conn.query(sql, values, (err, rs) => {
             try {
@@ -108,6 +96,7 @@ router.put("/:articleIdx", (req, res) => {
         })
     } catch (e) {
         result.message = e.message;
+        res.send(result);
     }
 })
 //게시글 삭제하기
@@ -121,8 +110,8 @@ router.delete("/:articleIdx", (req, res) => {
     }
     try {
         //if (req.session.idx !== userIdx) throw new Error("사용자 idx가 불일치")
-        const sql = "DELETE FROM article WHERE idx = ?";
-        const values = [articleIdx];
+        const sql = "DELETE FROM article WHERE idx = ? AND user_idx = ?";
+        const values = [articleIdx, userIdx];
         const conn = mysql.createConnection(dbconfig); // db연결
         conn.query(sql, values, (err, rs) => {
             try {
